week-9/froshims/requirements.txt
about-time==3.1.1
absl-py==1.2.0
alive-progress==2.4.1
apt-xapian-index==0.49
astroid==2.12.9
asttokens==2.0.8
autopep8==1.7.0
backcall==0.2.0
bandit==1.7.4
blinker==1.4
cachetools==5.2.0
certifi==2022.9.14
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
dill==0.3.5.1
distro==1.7.0
distro-info===1.1build1
entrypoints==0.4
executing==1.0.0
Flask==2.2.2
Flask-Mail==0.9.1
fuse-python==1.0.2
gitdb==4.0.9
GitPython==3.1.28
google-auth==2.11.1
google-auth-oauthlib==0.4.6
gpg===1.16.0-unknown
grapheme==0.6.0
grpcio==1.48.1
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
ipykernel==6.15.2
ipython==8.4.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.3.5
kazam==1.4.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.7.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
more-itertools==8.10.0
natsort==8.0.2
nest-asyncio==1.5.5
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.4.4
parso==0.8.3
pbr==5.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
progressbar==2.5
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.9.1
pycups==2.0.1
pydocstyle==6.1.1
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pylibacl==0.6.0
pylint==2.15.2
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
pytz==2022.2.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.2.1
reportlab==3.6.8
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
stack-data==0.5.0
stevedore==4.0.0
systemd-python==234
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
tornado==6.2
traitlets==5.3.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
